+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added new character order: Custom Order - Customize the order of your characters
 - You can now move all windows halfway off the screen
 - Focusing a window will now move it to the front of other windows